Move things around to create the Presentation project
