+++ MauiGeoQuiz.Presentation/Platforms/iOS/PlatformClass1.cs
@@ -1,0 +1,6 @@
+﻿namespace MauiGeoQuiz.Presentation;
+
+// All the code in this file is only included on iOS.
+public class PlatformClass1
+{
+}
