+++ MauiGeoQuiz.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35208.52
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MauiGeoQuiz", "MauiGeoQuiz\MauiGeoQuiz.csproj", "{CDE6521C-652F-4DD7-A573-E5F0C18D571C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MauiGeoQuiz", "MauiGeoQuiz\MauiGeoQuiz.csproj", "{CDE6521C-652F-4DD7-A573-E5F0C18D571C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MauiGeoQuiz.Presentation", "MauiGeoQuiz.Presentation\MauiGeoQuiz.Presentation.csproj", "{D89A76D7-A149-4686-8AED-4C263609D6B0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,6 +19,10 @@ Global
 		{CDE6521C-652F-4DD7-A573-E5F0C18D571C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CDE6521C-652F-4DD7-A573-E5F0C18D571C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{CDE6521C-652F-4DD7-A573-E5F0C18D571C}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{D89A76D7-A149-4686-8AED-4C263609D6B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D89A76D7-A149-4686-8AED-4C263609D6B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D89A76D7-A149-4686-8AED-4C263609D6B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D89A76D7-A149-4686-8AED-4C263609D6B0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
