+++ MauiGeoQuiz/MauiGeoQuiz.csproj
@@ -64,6 +64,10 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <ProjectReference Include="..\MauiGeoQuiz.Presentation\MauiGeoQuiz.Presentation.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
 	  <Compile Update="MainPage.xaml.cs">
 	    <DependentUpon>MainPage.xaml</DependentUpon>
 	  </Compile>
