+++ MauiGeoQuiz/Services/NavigationService.cs
@@ -1,4 +1,6 @@
-﻿namespace MauiGeoQuiz.Services;
+﻿using MauiGeoQuiz.Presentation.Services;
+
+namespace MauiGeoQuiz.Services;
 public class NavigationService : INavigationService
 {
     public async Task NavigateToPage(string page)
